backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
dotenv==0.9.9